+AntiAdManager
Added to Config:
Threshold:
    Block: 0.3
    ReduceThreshold: 0.1
    MaxLinkLength: 10
    Clear: true
    BlockFor: 30

Players get added an "error" in a add over message length error so long messages have less error even if there is an ad in there so we store the error and everytime someone writes something the error stacks if the error is > Block the players get muted for "BlockFor" seconds.

The MaxLinkLength: 10 means that a link has an error of 10 characters because everything around a "." in a range of 10 characters gets compressed if you increase that the error per message will increase if you decrease that the error per message get decreased. (Still in relation to the messafe length)

if you write now a long sentence like: "if you have somany friends get some other people on here.don't team with hackerman." you wont get punished anymore but if you write "Please concider joining the minehub(.)com community" you get only punished if you write it again (for the confidence) so we have close to no false positives but it takes more messages to block a player who advertises.

I think thats better, because now every Seerver owner can decide better when players get punished.
